+++ apps/sftp-source-rabbit/src/main/resources/application.properties
@@ -5,5 +5,5 @@ info.app.version=@project.version@
 management.endpoints.web.exposure.include=*
 #spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration
 #User added for convenience
-spring.cloud.stream.function.definition=transfer
+#spring.cloud.stream.function.definition=transfer
