+++ sftp-source-downloaders/sftp-source-downloader-cf-volume-services/src/main/java/org/springframework/cloud/stream/app/sftp/source/downloader/cf/volume/services/NFSInputStreamPersisterAutoConfiguration.java
@@ -16,16 +16,13 @@
 
 package org.springframework.cloud.stream.app.sftp.source.downloader.cf.volume.services;
 
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.cloud.stream.app.sftp.source.downloader.core.FileInputStreamPersister;
-import org.springframework.cloud.stream.app.sftp.source.downloader.core.InputStreamPersister;
-
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnCloudPlatform;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.cloud.CloudPlatform;
-import org.springframework.cloud.stream.app.sftp.source.downloader.core.SftpSourceDowloaderProperties;
+import org.springframework.cloud.stream.app.sftp.source.downloader.core.FileInputStreamPersister;
+import org.springframework.cloud.stream.app.sftp.source.downloader.core.InputStreamPersister;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.Environment;
@@ -38,7 +35,6 @@ import org.springframework.lang.Nullable;
  **/
 @Configuration
 @ConditionalOnCloudPlatform(CloudPlatform.CLOUD_FOUNDRY)
-@EnableConfigurationProperties(SftpSourceDowloaderProperties.class)
 @ConditionalOnProperty(value = "sftp.transfer-to", havingValue = "CF_VOLUME")
 public class NFSInputStreamPersisterAutoConfiguration {
 
@@ -71,7 +67,7 @@ public class NFSInputStreamPersisterAutoConfiguration {
 		public Object postProcessAfterInitialization(Object bean, String beanName) throws BeansException {
 			if (bean instanceof VcapService) {
 				bean = NFS.load(environment.getProperty("VCAP_SERVICES"),
-					environment.getProperty("nfs.service.name","nfs"));
+					environment.getProperty("nfs.service.name", "nfs"));
 			}
 			return bean;
 		}
