+++ sftp-source-downloaders/sftp-source-downloader-core/src/main/java/org/springframework/cloud/stream/app/sftp/source/downloader/core/SftpFunctionEnvironmentPostProcessor.java
@@ -16,26 +16,21 @@
 
 package org.springframework.cloud.stream.app.sftp.source.downloader.core;
 
-import org.springframework.boot.context.properties.ConfigurationProperties;
+import java.util.Collections;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.env.EnvironmentPostProcessor;
+import org.springframework.core.env.ConfigurableEnvironment;
+import org.springframework.core.env.MapPropertySource;
 
 /**
  * @author David Turanski
  **/
-@ConfigurationProperties(prefix = "sftp")
-public class SftpSourceDowloaderProperties {
-
-	public enum TransferType {LOCAL,CF_VOLUME,S3}
-
-	/**
-	 * The target storage to use for the file transfer (LOCAL, CF_VOLUME, S3) default is `LOCAL`
-	 */
-	private TransferType transferTo = TransferType.LOCAL;
-
-	public TransferType getTransferTo() {
-		return transferTo;
-	}
-
-	public void setTransferTo(TransferType transferTo) {
-		this.transferTo = transferTo;
+public class SftpFunctionEnvironmentPostProcessor implements EnvironmentPostProcessor {
+	@Override
+	public void postProcessEnvironment(ConfigurableEnvironment environment, SpringApplication application) {
+		environment.getPropertySources()
+			.addLast(new MapPropertySource("function-definitions",
+				Collections.singletonMap("spring.cloud.stream.function.definition", "transfer")));
 	}
 }
