--- sftp-source-downloaders/sftp-source-downloader-core/src/main/resources/META-INF/spring-configuration-metadata-whitelist.properties
@@ -1,2 +1,0 @@
-configuration-properties.classes=org.springframework.cloud.stream.app.sftp.source.downloader.core.SftpSourceDowloaderProperties
-
