+++ sftp-source-downloaders/sftp-source-downloader-core/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.springframework.cloud.stream.app.sftp.source.downloader.core.SftpFunctionAutoConfiguration
+org.springframework.boot.env.EnvironmentPostProcessor=org.springframework.cloud.stream.app.sftp.source.downloader.core.SftpFunctionEnvironmentPostProcessor
